Add a rudimentary ALSA-based mixer implementation.
This implementation doesn't work properly for all controls. Currently
only playback controls have been tested, and of them only those work
which have a joined playback volume control.

+++ src/server/remote-control.h
@@ -129,8 +129,8 @@ struct mixer;
 
 int mixer_create(struct mixer **mixerp);
 int mixer_free(struct mixer *mixer);
-int mixer_set_volume(struct mixer *mixer, unsigned short control, uint8_t volume);
-int mixer_get_volume(struct mixer *mixer, unsigned short control, uint8_t *volumep);
+int mixer_set_volume(struct mixer *mixer, unsigned short control, unsigned int volume);
+int mixer_get_volume(struct mixer *mixer, unsigned short control, unsigned int *volumep);
 int mixer_set_mute(struct mixer *mixer, unsigned short control, bool mute);
 int mixer_is_muted(struct mixer *mixer, unsigned short control, bool *mutep);
